Eliminated some gcc warnings

+++ Emulator/Foundation/Aliases.h
@@ -11,20 +11,25 @@
 #define _ALIASES_H
 
 #include <stdint.h>
+#include <stddef.h>
 #include <assert.h>
 
 //
 // Basic types
 //
 
-typedef int8_t   i8;
-typedef int16_t  i16;
-typedef int32_t  i32;
-typedef int64_t  i64;
-typedef uint8_t  u8;
-typedef uint16_t u16;
-typedef uint32_t u32;
-typedef uint64_t u64;
+/* Note: We don't map i64 to int64_t, because this type is mapped to long long
+ * on macOS, but mapped to long on Linux. This would require different format
+ * strings when such a value is printed via printf().
+ */
+typedef int8_t             i8;
+typedef int16_t            i16;
+typedef int32_t            i32;
+typedef long long          i64;
+typedef uint8_t            u8;
+typedef uint16_t           u16;
+typedef uint32_t           u32;
+typedef unsigned long long u64;
 
 
 //
